Set up Robolectric and add one test to test MainActivity. Code refactor.
