+++ app/build.gradle
@@ -11,8 +11,8 @@ android {
     buildToolsVersion "29.0.1"
     defaultConfig {
         applicationId "com.example.newsFeed"
-        minSdkVersion 29
-        targetSdkVersion 29
+        minSdkVersion 28
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -32,6 +32,12 @@ android {
         }
 
         kotlinOptions { jvmTarget = '1.8' }
+
+        testOptions {
+            unitTests {
+                includeAndroidResources = true
+            }
+        }
     }
     buildTypes {
         debug {
@@ -94,6 +100,7 @@ dependencies {
 
     // Test
     testImplementation 'junit:junit:4.12'
+    testImplementation 'org.robolectric:robolectric:4.4'
     androidTestImplementation 'androidx.test:runner:1.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 }
