+++ app/src/main/java/com/example/newsFeed/mainFeed/domain/NoteUseCase.kt
@@ -2,7 +2,7 @@ package com.example.newsFeed.mainFeed.domain
 
 import android.util.Log
 import com.example.newsFeed.mainFeed.model.Note
-import com.example.newsFeed.mainFeed.presentation.MainFeedViewModel
+import com.example.newsFeed.mainFeed.presentation.Change
 import com.example.newsFeed.mainFeed.repository.NoteRepository
 import javax.inject.Inject
 
@@ -10,10 +10,10 @@ class NoteUseCase @Inject constructor(private val repository: NoteRepository) {
 
     //For the purpose of MVI, a suspend function returning a list of notes is probably better than
     // a Flow of list of notes which receives the updates to the notes automatically.
-    suspend fun retrieveAllNotes(): MainFeedViewModel.Change.AllNotesRetrieved
-        = MainFeedViewModel.Change.AllNotesRetrieved(repository.retrieveAllNotes())
+    suspend fun retrieveAllNotes(): Change.AllNotesRetrieved
+        = Change.AllNotesRetrieved(repository.retrieveAllNotes())
 
-    suspend fun insertNote(note: Note): MainFeedViewModel.Change {
+    suspend fun insertNote(note: Note): Change {
         var newNote = note.copy()
         if (newNote.imageUri.isEmpty()) {
             val imageUrl = repository.getRandomCatImage().imageUrl
@@ -23,12 +23,12 @@ class NoteUseCase @Inject constructor(private val repository: NoteRepository) {
         Log.d(this.javaClass.simpleName, "Attempt to insert note $note")
         repository.insertNode(newNote)
 
-        return MainFeedViewModel.Change.NoteInserted(repository.retrieveAllNotes())
+        return Change.NoteInserted(repository.retrieveAllNotes())
     }
 
-    suspend fun deleteNote(note: Note): MainFeedViewModel.Change.NoteDeleted {
+    suspend fun deleteNote(note: Note): Change.NoteDeleted {
         repository.deleteNote(note)
-        return MainFeedViewModel.Change.NoteDeleted(repository.retrieveAllNotes())
+        return Change.NoteDeleted(repository.retrieveAllNotes())
     }
 
     // TODO Update one Note
