+++ app/src/main/java/com/example/newsFeed/mainFeed/presentation/State.kt
@@ -1,0 +1,5 @@
+package com.example.newsFeed.mainFeed.presentation
+
+import com.example.newsFeed.mainFeed.model.Note
+
+data class State(val isLoading: Boolean, val notes: List<Note> = emptyList(), val error: Throwable? = null)
